+++ tests/integration/pipelex/plugins/test_mistral.py
@@ -4,7 +4,7 @@ from rich.console import Console
 from rich.table import Table
 
 from pipelex import pretty_print
-from pipelex.plugins.mistral.mistral_llms import list_mistral_models
+from pipelex.plugins.mistral.mistral_llms import mistral_list_available_models
 from pipelex.tools.environment import all_env_vars_are_set, any_env_var_is_placeholder
 
 REQUIRED_ENV_VARS = ["MISTRAL_API_KEY"]
@@ -21,7 +21,7 @@ class TestMistral:
             pytest.skip(f"Some key(s) missing amongst {REQUIRED_ENV_VARS}")
         if any_env_var_is_placeholder(REQUIRED_ENV_VARS):
             pytest.skip(f"Some key(s) among {REQUIRED_ENV_VARS} are a placeholder, can't be used to test listing models")
-        mistral_models_list = list_mistral_models()
+        mistral_models_list = mistral_list_available_models()
         if pytestconfig.get_verbosity() >= 2:
             # Create and configure the table
             console = Console()
@@ -53,7 +53,7 @@ class TestMistral:
             pytest.skip(f"Some key(s) missing amongst {REQUIRED_ENV_VARS}")
         if any_env_var_is_placeholder(REQUIRED_ENV_VARS):
             pytest.skip(f"Some key(s) among {REQUIRED_ENV_VARS} are a placeholder, can't be used to test listing models")
-        mistral_models_list = list_mistral_models()
+        mistral_models_list = mistral_list_available_models()
         mistral_model_ids = [{"id": model.id, "aliases": model.aliases} for model in mistral_models_list]
         if pytestconfig.get_verbosity() >= 2:
             pretty_print(mistral_model_ids)